strset: do not put a pointer to a local variable to nlctx
Build with gcc12 complains that strset_reply_cb() puts a pointer to a local
function into nlctx->devname and leaves it there on function exit. As the
only reason for this assignment is a call to dev_ok() helper shortly after
this, open code the check instead to avoid leaving pointer to a variable
with function scope behind. We can also move the check insid the if block
as ifindex cannot be non-zero otherwise.

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>