ethtool: fec: Change the prompt string to adapt to current situations
Majority upstream drivers uses `Configured FEC encodings` to report
supported modes. At which point it is better to change the text in
ethtool user space that changes the meaning of the field, which is
better to suit for the current situations.

So changing `Configured FEC encodings` to `Supported/Configured FEC
encodings` to adapt to both implementations.

Signed-off-by: Yu Xiao <yu.xiao@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>